+++ lib/ui/components/shimmer.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:shimmer/shimmer.dart';
 
+import '../../state.dart';
+import '../common/app_colors.dart';
+
 Widget buildShimmerServiceItem(BuildContext context) {
   final bool isDarkMode = Theme.of(context).brightness == Brightness.dark;
 
@@ -53,6 +56,7 @@ Widget buildShimmerServiceItem(BuildContext context) {
     ),
   );
 }
+
 Widget buildTagsShimmerLoading(bool isSingleRow) {
   return Shimmer.fromColors(
     baseColor: Colors.grey[300]!,
@@ -99,6 +103,63 @@ Widget buildTagsShimmerLoading(bool isSingleRow) {
             ),
           );
         },
+      ),
+    ),
+  );
+}
+
+Widget buildShimmerContainer() {
+  return Shimmer.fromColors(
+    baseColor: uiMode.value == AppUiModes.dark
+        ? Colors.grey[700]!
+        : Colors.grey[300]!,
+    highlightColor: uiMode.value == AppUiModes.dark
+        ? Colors.grey[300]!
+        : Colors.grey[100]!,
+    child: Container(
+      width: double.infinity,
+      height: 200,
+      decoration: BoxDecoration(
+        color: kcSecondaryColor,
+        borderRadius: BorderRadius.circular(20),
+      ),
+    ),
+  );
+}
+
+Widget buildShimmerQuickActions() {
+  return Shimmer.fromColors(
+    baseColor: uiMode.value == AppUiModes.dark
+        ? Colors.grey[700]!
+        : Colors.grey[300]!,
+    highlightColor: uiMode.value == AppUiModes.dark
+        ? Colors.grey[300]!
+        : Colors.grey[100]!,
+    child: Container(
+      width: double.infinity,
+      height: 60,
+      decoration: BoxDecoration(
+        color: Colors.grey[300],
+        borderRadius: BorderRadius.circular(10),
+      ),
+    ),
+  );
+}
+
+Widget buildShimmerSlider() {
+  return Shimmer.fromColors(
+    baseColor: uiMode.value == AppUiModes.dark
+        ? Colors.grey[700]!
+        : Colors.grey[300]!,
+    highlightColor: uiMode.value == AppUiModes.dark
+        ? Colors.grey[300]!
+        : Colors.grey[100]!,
+    child: Container(
+      height: 300,
+      width: double.infinity,
+      decoration: BoxDecoration(
+        color: Colors.grey[300],
+        borderRadius: BorderRadius.circular(10),
       ),
     ),
   );
