+++ lib/ui/views/dashboard/dashboard_viewmodel.dart
@@ -28,7 +28,7 @@ class DashboardViewModel extends BaseViewModel {
   List<ProjectResource> projectResources = [];
   List<Raffle> featuredRaffle = [];
   List<Product> productList = [];
-  List<String> brands = [];
+  List<String> brands = []; // ← stable list; built from full productList only
   List<Product> filteredProductList = [];
   List<Category> filteredCategories = [];
   List<Category> categories = [];
@@ -76,57 +76,39 @@ class DashboardViewModel extends BaseViewModel {
   }
 
   void _applyFilters() {
-    // Create a Set to track unique product IDs and prevent duplicates
-    Set<String> seenProductIds = {};
-    List<Product> filtered = [];
+    final seenProductIds = <String>{};
+    final filtered = <Product>[];
 
-    // Start with all products
-    for (Product product in productList) {
-      // Skip if we've already seen this product ID
-      if (seenProductIds.contains(product.id)) {
-        continue;
-      }
+    for (final product in productList) {
+      // skip duplicates
+      final pid = product.id;
+      if (pid == null || seenProductIds.contains(pid)) continue;
 
-      bool matchesFilters = true;
+      bool matches = true;
 
-      // Apply category filter
+      // category filter
       if (selectedId != allCategoriesId) {
-        matchesFilters = matchesFilters && (product.categoryId == selectedId);
+        matches = matches && (product.categoryId == selectedId);
       }
 
-      // Apply brand filter
+      // brand filter
       if (selectedBrand.isNotEmpty) {
-        matchesFilters = matchesFilters && (product.brandName == selectedBrand);
+        matches = matches && (product.brandName == selectedBrand);
       }
 
-      // Apply tag filter
+      // tag filter
       if (_selectedTag != null) {
-        bool hasMatchingTag = false;
-        if (product.tags != null && product.tags!.isNotEmpty) {
-          hasMatchingTag = product.tags!.any((tag) => tag.id == _selectedTag!.id);
-        }
-        matchesFilters = matchesFilters && hasMatchingTag;
+        final hasTag = (product.tags ?? []).any((t) => t.id == _selectedTag!.id);
+        matches = matches && hasTag;
       }
 
-      // Add to filtered list if it matches all filters and hasn't been seen
-      if (matchesFilters) {
+      if (matches) {
         filtered.add(product);
-        seenProductIds.add(product.id!);
+        seenProductIds.add(pid);
       }
     }
 
     filteredProductList = filtered;
-    print('After applying filters: ${filteredProductList.length} unique products');
-
-    // Update brands based on filtered products
-    Set<String> uniqueBrands = {};
-    for (Product product in filteredProductList) {
-      if (product.brandName != null && product.brandName!.isNotEmpty) {
-        uniqueBrands.add(product.brandName!);
-      }
-    }
-    brands = uniqueBrands.toList();
-
     notifyListeners();
   }
 
@@ -136,23 +118,25 @@ class DashboardViewModel extends BaseViewModel {
     _selectedTag = null;
 
     // Remove duplicates from productList before applying to filtered list
-    Set<String> seenIds = {};
-    List<Product> uniqueProducts = [];
-    for (Product product in productList) {
-      if (!seenIds.contains(product.id)) {
+    final seenIds = <String>{};
+    final uniqueProducts = <Product>[];
+    for (final product in productList) {
+      if (product.id != null && !seenIds.contains(product.id)) {
         uniqueProducts.add(product);
         seenIds.add(product.id!);
       }
     }
 
     filteredProductList = uniqueProducts;
-    Set<String> uniqueBrands = {};
-    for (Product product in filteredProductList) {
-      if (product.brandName != null && product.brandName!.isNotEmpty) {
-        uniqueBrands.add(product.brandName!);
-      }
+
+    // Build brands from the FULL catalog (not filtered)
+    final uniqueBrands = <String>{};
+    for (final product in productList) {
+      final b = product.brandName;
+      if (b != null && b.isNotEmpty) uniqueBrands.add(b);
     }
-    brands = uniqueBrands.toList();
+    brands = uniqueBrands.toList()..sort();
+
     notifyListeners();
   }
 
@@ -160,8 +144,9 @@ class DashboardViewModel extends BaseViewModel {
     _selectedTag = null;
     selectedId = id;
     _applyFilters();
-    brands = filteredProductList.map((product) => product.brandName ?? '').toSet().toList();
-    brands.removeWhere((brand) => brand.isEmpty);
+
+    // ⚠️ Do NOT rebuild `brands` from `filteredProductList`.
+    // Keep `brands` as a stable list from the full catalog so other chips remain visible.
   }
 
   void setSelectedBrand(String brand) {
@@ -171,8 +156,14 @@ class DashboardViewModel extends BaseViewModel {
   }
 
   void setSelectedTag(Tag? tag) {
+    clearAllFilters();
     _selectedTag = tag;
-    // When a tag is selected, apply filters immediately
+    final uniqueBrands = <String>{};
+    for (final product in productList) {
+      final b = product.brandName;
+      if (b != null && b.isNotEmpty) uniqueBrands.add(b);
+    }
+    brands = uniqueBrands.toList()..sort();
     _applyFilters();
   }
 
@@ -242,10 +233,10 @@ class DashboardViewModel extends BaseViewModel {
             .toList();
 
         // Remove duplicates from loaded products
-        Set<String> seenIds = {};
+        final seenIds = <String>{};
         productList = [];
-        for (Product product in loadedProducts) {
-          if (!seenIds.contains(product.id)) {
+        for (final product in loadedProducts) {
+          if (product.id != null && !seenIds.contains(product.id)) {
             productList.add(product);
             seenIds.add(product.id!);
           }
@@ -254,13 +245,14 @@ class DashboardViewModel extends BaseViewModel {
         filteredProductList = List.from(productList);
         print('loaded ${productList.length} unique products from local storage');
 
-        Set<String> uniqueBrands = {};
-        for (Product product in filteredProductList) {
-          if (product.brandName != null && product.brandName!.isNotEmpty) {
-            uniqueBrands.add(product.brandName!);
-          }
+        // Build brands from FULL productList (stable)
+        final uniqueBrands = <String>{};
+        for (final product in productList) {
+          final b = product.brandName;
+          if (b != null && b.isNotEmpty) uniqueBrands.add(b);
         }
-        brands = uniqueBrands.toList();
+        brands = uniqueBrands.toList()..sort();
+
         rebuildUi();
       } else {
         print('no value to load');
@@ -329,9 +321,9 @@ class DashboardViewModel extends BaseViewModel {
         final totalPages = res.data["pagination"]["totalPages"];
 
         // Add new products, avoiding duplicates
-        Set<String> existingIds = productList.map((p) => p.id!).toSet();
-        for (Product newProduct in newProducts) {
-          if (!existingIds.contains(newProduct.id)) {
+        final existingIds = productList.map((p) => p.id!).toSet();
+        for (final newProduct in newProducts) {
+          if (newProduct.id != null && !existingIds.contains(newProduct.id)) {
             productList.add(newProduct);
           }
         }
@@ -339,13 +331,13 @@ class DashboardViewModel extends BaseViewModel {
         // Apply current filters to all products
         _applyCurrentFilters();
 
-        Set<String> uniqueBrands = {};
-        for (Product product in productList) {
-          if (product.brandName != null && product.brandName!.isNotEmpty) {
-            uniqueBrands.add(product.brandName!);
-          }
+        // Rebuild brands from FULL productList (stable)
+        final uniqueBrands = <String>{};
+        for (final product in productList) {
+          final b = product.brandName;
+          if (b != null && b.isNotEmpty) uniqueBrands.add(b);
         }
-        brands = uniqueBrands.toList();
+        brands = uniqueBrands.toList()..sort();
 
         if (currentPage >= totalPages) {
           isLastPage = true;
@@ -354,10 +346,10 @@ class DashboardViewModel extends BaseViewModel {
         }
 
         // Save unique products to storage
-        Set<String> seenIds = {};
-        List<Product> uniqueProducts = [];
-        for (Product product in productList) {
-          if (!seenIds.contains(product.id)) {
+        final seenIds = <String>{};
+        final uniqueProducts = <Product>[];
+        for (final product in productList) {
+          if (product.id != null && !seenIds.contains(product.id)) {
             uniqueProducts.add(product);
             seenIds.add(product.id!);
           }
@@ -380,12 +372,12 @@ class DashboardViewModel extends BaseViewModel {
 
   void _applyCurrentFilters() {
     // Create a Set to track unique product IDs and prevent duplicates
-    Set<String> seenProductIds = {};
-    List<Product> filtered = [];
+    final seenProductIds = <String>{};
+    final filtered = <Product>[];
 
-    for (Product product in productList) {
+    for (final product in productList) {
       // Skip if we've already seen this product ID
-      if (seenProductIds.contains(product.id)) {
+      if (product.id != null && seenProductIds.contains(product.id)) {
         continue;
       }
 
@@ -413,7 +405,9 @@ class DashboardViewModel extends BaseViewModel {
       // Add to filtered list if it matches all filters and hasn't been seen
       if (matchesFilters) {
         filtered.add(product);
-        seenProductIds.add(product.id!);
+        if (product.id != null) {
+          seenProductIds.add(product.id!);
+        }
       }
     }
