+++ lib/ui/views/dashboard/widgets/app_bar_actions.dart
@@ -1,0 +1,67 @@
+
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+Widget actionContainer(String imagePath, String title, BuildContext context) {
+  return Padding(
+    padding: const EdgeInsets.only(left: 0.0, right: 8.0),
+    child: ClipRRect(
+      borderRadius: BorderRadius.circular(10.0),
+      child: Stack(
+        children: [
+          Container(
+            width: MediaQuery.of(context).size.width,
+            height: MediaQuery.of(context).size.height,
+            decoration: BoxDecoration(
+              color: Colors.grey[200],
+              boxShadow: const [
+                BoxShadow(
+                  color: Colors.black12,
+                  blurRadius: 5.0,
+                  spreadRadius: 1.0,
+                  offset: Offset(0, 3),
+                ),
+              ],
+            ),
+            child: Image.asset(
+              imagePath,
+              fit: BoxFit.cover,
+            ),
+          ),
+          // Overlay
+          Positioned.fill(
+            child: Container(
+              color:
+              Colors.black.withOpacity(0.5),
+            ),
+          ),
+          // Title Text
+          Positioned(
+            bottom: 8,
+            left: 8,
+            right: 8,
+            child: Text(
+              title,
+              style: const TextStyle(
+                color: Colors.white,
+                fontWeight: FontWeight.bold,
+                fontSize: 12,
+                shadows: [
+                  Shadow(
+                    blurRadius: 4.0,
+                    color: Colors.black,
+                    offset: Offset(0, 2),
+                  ),
+                ],
+              ),
+              textAlign: TextAlign.center,
+              maxLines: 2,
+              overflow: TextOverflow.ellipsis,
+            ),
+          ),
+        ],
+      ),
+    ),
+  );
+}
